requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
dnspython==2.6.1
mongoengine==0.28.0
pillow==10.2.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
setuptools==68.2.2
six==1.16.0
sqlparse @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/sqlparse_1699544474746/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
wheel==0.41.2
jieba~=0.42.1
beautifulsoup4~=4.12.3